+++ apps/backend/services/custom/containers.py
@@ -1,9 +1,9 @@
 import datetime
 
+from backend.database import database
+from backend.mapper_decorator import apply_mapper
 from sqlalchemy import select
 
-from src.database import database
-from src.mapper_decorator import apply_mapper
 from .mappers.containers import map_containers
 from .models.container import ContainerModel
