+++ apps/backend/services/custom/models/container.py
@@ -1,9 +1,8 @@
 import enum
 
+from backend.database import Base
 from sqlalchemy import Enum, String, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column
-
-from src.database import Base
 
 
 class ContainerType(enum.Enum):
