+++ apps/backend/services/custom/points.py
@@ -1,10 +1,10 @@
 import datetime
 from functools import partial
 
+from backend.database import database
+from backend.mapper_decorator import apply_mapper
 from sqlalchemy import select
 
-from src.database import database
-from src.mapper_decorator import apply_mapper
 from .mappers.points import map_points
 from .models import CompanyModel, PointModel, RailRouteModel, SeaRouteModel
