+++ apps/backend/services/fesco/containers.py
@@ -2,8 +2,8 @@ import datetime
 import os
 
 import aiohttp
+from backend.mapper_decorator import apply_mapper
 
-from src.mapper_decorator import apply_mapper
 from .mappers.containers import map_containers
