+++ apps/backend/services/fesco/routes.py
@@ -3,8 +3,8 @@ import datetime
 import os
 
 import aiohttp
+from backend.mapper_decorator import apply_mapper
 
-from src.mapper_decorator import apply_mapper
 from .mappers.routes import map_routes
