+++ apps/backend/utils/string_formatters.py
