+++ apps/frontend/res/js/alerts.js
