+++ apps/frontend/res/js/rates.js
